feat: component cleanup
